feat(vsc): implement a activate function
